+++ cancel_ride.php
@@ -12,9 +12,11 @@ $conn = mysqli_connect($servername, $username, $password,"carpooldb");
 if (!$conn) {
     die("Connection failed: " . mysqli_connect_error());
 }
-
-$query1="delete from booked where jid = ".$_SESSION['cj_id'];
-$query2="update journey set sel = 1 where j_id = ".$_SESSION['cj_id'];
+$query3="select seats from booked where jid = ".$_GET['cj_id'];
+$result = $conn->query($query3);
+$row=$result->fetch_assoc();
+$query1="delete from booked where jid = ".$_GET['cj_id'];
+$query2="update journey set sel = 1, seats = seats + ".$row['seats']." where j_id = ".$_GET['cj_id'];
 $result = $conn->query($query1);
 $result = $conn->query($query2);
 header("Location:profile.php");
