--- titl/css/colors/green.css
@@ -1,117 +1,0 @@
-a{
-	color: #94C523;
-}
-a:hover{
-	color: #94C523;
-}
-.btn-common{
-	border-color: #94C523;
-    background-color: #94C523;
-}
-.btn-common:hover{
-	border-color: #94C523;
-    color: #94C523;
-}
-.btn-border:hover{
-	background-color: #94C523;
-}
-#subscribe .sub_btn:hover{
-	background: #a5da2d;
-}
-.section-header .section-title span{
-	color: #94C523;
-}
-.section-header .lines{
-	border-color: #94C523; 
-}
-.section-header .lines:before{
-	background: #94C523;
-}
-.overlay{
-	background: #94C523;
-}
-#hero-area .overlay{
-	background: #94C523;
-}
-.navbar-light .navbar-nav .active > .nav-link, .navbar-light .navbar-nav .nav-link.active, .navbar-light .navbar-nav .nav-link.active::before, .navbar-light .navbar-nav .nav-link.open, .navbar-light .navbar-nav .open > .nav-link{
-	color: #94C523;
-}
-.navbar-light .navbar-nav .nav-link:focus, .navbar-light .navbar-nav .nav-link:hover, .navbar-light .navbar-nav .nav-link:hover:before{
-	color: #94C523;
-}
-.navbar-light .navbar-nav .nav-link:before{
-	background: #94C523;
-}
-.menu-wrap .navbar-nav li .active,
-.menu-wrap .navbar-nav li .active::before{
-	color: #94C523;
-}
-.menu-wrap a:hover, .menu-wrap a:focus{
-	color: #94C523;
-}
-.icon-list a::before {
-    background: #94C523;
-}
-#about .about-boxes .icon{
-	border-color: #94C523;
-}
-#about .about-boxes .icon i{
-	color: #94C523;
-}
-.item-boxes .icon{
-	border-color: #94C523;
-}
-.item-boxes .icon i{
-	color: #94C523;
-}
-#features .box-item .icon{
-	border-color: #94C523;
-}
-#features .box-item .icon i{
-	color: #94C523;
-}
-#features .box-item:hover .icon{
-	background: #94C523;
-}
-.video-promo .video-promo-content .video-popup:hover{
-	background: #94C523;
-	border-color: #94C523;
-}
-.shot-item .overlay:before, 
-.shot-item .overlay:after{
-	background: #94C523;
-}
-.controls .mixitup-control-active{
-	color: #94C523!important;
-	border-color: #94C523;
-}
-.owl-theme .owl-controls .owl-page span{
-	background: #94C523;
-}
-.pricing-table .icon{
-	color: #94C523;
-	border-color: #94C523;
-}
-.pricing-table .pricing-details span{
-	color: #94C523;
-}
-#blog .blog-item-img:before {
-	background: rgba(148, 197, 35, 0.7);
-}
-.form-control:focus{
-	box-shadow: 0 0 2px #94C523;
-}
-#subscribe .sub_btn{
-	background: #94C523;
-}
-#contact:before{
-	background: #94C523;
-}
-.back-to-top i{
-	background-color: #94C523;
-}
-
-.tp-bullets.preview4 .bullet:hover, .tp-bullets.preview4 .bullet.selected{
-	border-color: #94C523!important;
-}
-
