--- titl/css/colors/lightgreen.css
@@ -1,117 +1,0 @@
-a{
-	color: #1EB858;
-}
-a:hover{
-	color: #1EB858;
-}
-.btn-common{
-	border-color: #1EB858;
-    background-color: #1EB858;
-}
-.btn-common:hover{
-	border-color: #1EB858;
-    color: #1EB858;
-}
-.btn-border:hover{
-	background-color: #1EB858;
-}
-#subscribe .sub_btn:hover{
-	background: #2ec969;
-}
-.section-header .section-title span{
-	color: #1EB858;
-}
-.section-header .lines{
-	border-color: #1EB858; 
-}
-.section-header .lines:before{
-	background: #1EB858;
-}
-.overlay{
-	background: #1EB858;
-}
-#hero-area .overlay{
-	background: #1EB858;
-}
-.navbar-light .navbar-nav .active > .nav-link, .navbar-light .navbar-nav .nav-link.active, .navbar-light .navbar-nav .nav-link.active::before, .navbar-light .navbar-nav .nav-link.open, .navbar-light .navbar-nav .open > .nav-link{
-	color: #1EB858;
-}
-.navbar-light .navbar-nav .nav-link:focus, .navbar-light .navbar-nav .nav-link:hover, .navbar-light .navbar-nav .nav-link:hover:before{
-	color: #1EB858;
-}
-.navbar-light .navbar-nav .nav-link:before{
-	background: #1EB858;
-}
-.menu-wrap .navbar-nav li .active,
-.menu-wrap .navbar-nav li .active::before{
-	color: #1EB858;
-}
-.menu-wrap a:hover, .menu-wrap a:focus{
-	color: #1EB858;
-}
-.icon-list a::before {
-    background: #1EB858;
-}
-#about .about-boxes .icon{
-	border-color: #1EB858;
-}
-#about .about-boxes .icon i{
-	color: #1EB858;
-}
-.item-boxes .icon{
-	border-color: #1EB858;
-}
-.item-boxes .icon i{
-	color: #1EB858;
-}
-#features .box-item .icon{
-	border-color: #1EB858;
-}
-#features .box-item .icon i{
-	color: #1EB858;
-}
-#features .box-item:hover .icon{
-	background: #1EB858;
-}
-.video-promo .video-promo-content .video-popup:hover{
-	background: #1EB858;
-	border-color: #1EB858;
-}
-.shot-item .overlay:before, 
-.shot-item .overlay:after{
-	background: #1EB858;
-}
-.controls .mixitup-control-active{
-	color: #1EB858!important;
-	border-color: #1EB858;
-}
-.owl-theme .owl-controls .owl-page span{
-	background: #1EB858;
-}
-.pricing-table .icon{
-	color: #1EB858;
-	border-color: #1EB858;
-}
-.pricing-table .pricing-details span{
-	color: #1EB858;
-}
-#blog .blog-item-img:before {
-	background: rgba(30, 184, 88, 0.7);
-}
-.form-control:focus{
-	box-shadow: 0 0 2px #1EB858;
-}
-#subscribe .sub_btn{
-	background: #1EB858;
-}
-#contact:before{
-	background: #1EB858;
-}
-.back-to-top i{
-	background-color: #1EB858;
-}
-
-.tp-bullets.preview4 .bullet:hover, .tp-bullets.preview4 .bullet.selected{
-	border-color: #1EB858!important;
-}
-
