--- titl/css/colors/orange.css
@@ -1,116 +1,0 @@
-a{
-	color: #F57D41;
-}
-a:hover{
-	color: #F57D41;
-}
-.btn-common{
-	border-color: #F57D41;
-    background-color: #F57D41;
-}
-.btn-common:hover{
-	border-color: #F57D41;
-    color: #F57D41;
-}
-.btn-border:hover{
-	background-color: #F57D41;
-}
-#subscribe .sub_btn:hover{
-	background: #ec6e2f;
-}
-.section-header .section-title span{
-	color: #F57D41;
-}
-.section-header .lines{
-	border-color: #F57D41; 
-}
-.section-header .lines:before{
-	background: #F57D41;
-}
-.overlay{
-	background: #F57D41;
-}
-#hero-area .overlay{
-	background: #F57D41;
-}
-.navbar-light .navbar-nav .active > .nav-link, .navbar-light .navbar-nav .nav-link.active, .navbar-light .navbar-nav .nav-link.active::before, .navbar-light .navbar-nav .nav-link.open, .navbar-light .navbar-nav .open > .nav-link{
-	color: #F57D41;
-}
-.navbar-light .navbar-nav .nav-link:focus, .navbar-light .navbar-nav .nav-link:hover, .navbar-light .navbar-nav .nav-link:hover:before{
-	color: #F57D41;
-}
-.navbar-light .navbar-nav .nav-link:before{
-	background: #F57D41;
-}
-.menu-wrap .navbar-nav li .active,
-.menu-wrap .navbar-nav li .active::before{
-	color: #F57D41;
-}
-.menu-wrap a:hover, .menu-wrap a:focus{
-	color: #F57D41;
-}
-.icon-list a::before {
-    background: #F57D41;
-}
-#about .about-boxes .icon{
-	border-color: #F57D41;
-}
-#about .about-boxes .icon i{
-	color: #F57D41;
-}
-.item-boxes .icon{
-	border-color: #F57D41;
-}
-.item-boxes .icon i{
-	color: #F57D41;
-}
-#features .box-item .icon{
-	border-color: #F57D41;
-}
-#features .box-item .icon i{
-	color: #F57D41;
-}
-#features .box-item:hover .icon{
-	background: #F57D41;
-}
-.video-promo .video-promo-content .video-popup:hover{
-	background: #F57D41;
-	border-color: #F57D41;
-}
-.shot-item .overlay:before, 
-.shot-item .overlay:after{
-	background: #F57D41;
-}
-.controls .mixitup-control-active{
-	color: #F57D41!important;
-	border-color: #F57D41;
-}
-.owl-theme .owl-controls .owl-page span{
-	background: #F57D41;
-}
-.pricing-table .icon{
-	color: #F57D41;
-	border-color: #F57D41;
-}
-.pricing-table .pricing-details span{
-	color: #F57D41;
-}
-#blog .blog-item-img:before {
-	background: rgba(245, 125, 65, 0.7);
-}
-.form-control:focus{
-	box-shadow: 0 0 2px #F57D41;
-}
-#subscribe .sub_btn{
-	background: #F57D41;
-}
-#contact:before{
-	background: #F57D41;
-}
-.back-to-top i{
-	background-color: #F57D41;
-}
-
-.tp-bullets.preview4 .bullet:hover, .tp-bullets.preview4 .bullet.selected{
-	border-color: #F57D41!important;
-}
