--- titl/css/colors/pink.css
@@ -1,117 +1,0 @@
-a{
-	color: #6B58CD;
-}
-a:hover{
-	color: #6B58CD;
-}
-.btn-common{
-	border-color: #6B58CD;
-    background-color: #6B58CD;
-}
-.btn-common:hover{
-	border-color: #6B58CD;
-    color: #6B58CD;
-}
-.btn-border:hover{
-	background-color: #6B58CD;
-}
-#subscribe .sub_btn:hover{
-	background: #543dcc;
-}
-.section-header .section-title span{
-	color: #6B58CD;
-}
-.section-header .lines{
-	border-color: #6B58CD; 
-}
-.section-header .lines:before{
-	background: #6B58CD;
-}
-.overlay{
-	background: #6B58CD;
-}
-#hero-area .overlay{
-	background: #6B58CD;
-}
-.navbar-light .navbar-nav .active > .nav-link, .navbar-light .navbar-nav .nav-link.active, .navbar-light .navbar-nav .nav-link.active::before, .navbar-light .navbar-nav .nav-link.open, .navbar-light .navbar-nav .open > .nav-link{
-	color: #6B58CD;
-}
-.navbar-light .navbar-nav .nav-link:focus, .navbar-light .navbar-nav .nav-link:hover, .navbar-light .navbar-nav .nav-link:hover:before{
-	color: #6B58CD;
-}
-.navbar-light .navbar-nav .nav-link:before{
-	background: #6B58CD;
-}
-.menu-wrap .navbar-nav li .active,
-.menu-wrap .navbar-nav li .active::before{
-	color: #6B58CD;
-}
-.menu-wrap a:hover, .menu-wrap a:focus{
-	color: #6B58CD;
-}
-.icon-list a::before {
-    background: #6B58CD;
-}
-#about .about-boxes .icon{
-	border-color: #6B58CD;
-}
-#about .about-boxes .icon i{
-	color: #6B58CD;
-}
-.item-boxes .icon{
-	border-color: #6B58CD;
-}
-.item-boxes .icon i{
-	color: #6B58CD;
-}
-#features .box-item .icon{
-	border-color: #6B58CD;
-}
-#features .box-item .icon i{
-	color: #6B58CD;
-}
-#features .box-item:hover .icon{
-	background: #6B58CD;
-}
-.video-promo .video-promo-content .video-popup:hover{
-	background: #6B58CD;
-	border-color: #6B58CD;
-}
-.shot-item .overlay:before, 
-.shot-item .overlay:after{
-	background: #6B58CD;
-}
-.controls .mixitup-control-active{
-	color: #6B58CD!important;
-	border-color: #6B58CD;
-}
-.owl-theme .owl-controls .owl-page span{
-	background: #6B58CD;
-}
-.pricing-table .icon{
-	color: #6B58CD;
-	border-color: #6B58CD;
-}
-.pricing-table .pricing-details span{
-	color: #6B58CD;
-}
-#blog .blog-item-img:before {
-	background: rgba(107, 88, 205, 0.7);
-}
-.form-control:focus{
-	box-shadow: 0 0 2px #6B58CD;
-}
-#subscribe .sub_btn{
-	background: #6B58CD;
-}
-#contact:before{
-	background: #6B58CD;
-}
-.back-to-top i{
-	background-color: #6B58CD;
-}
-
-.tp-bullets.preview4 .bullet:hover, .tp-bullets.preview4 .bullet.selected{
-	border-color: #6B58CD!important;
-}
-
